+++ src/util/index.ts
@@ -1,1 +1,1 @@
-export * from "./file/index.js"
+export * from "../file/index.js"